# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Counselling Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Counselling {
  id: string;
  name: string;
  slug: string;
  /** e.g. Engineering, Architecture */
  type: string;
  /** e.g. National, State */
  level: string;
  states: string[];
  startMonth: string;
  endMonth: string;
  startYear: number;
  /** @nullable */
  logoUrl: string | null;
  description: string;
  /** @nullable */
  totalColleges?: number | null;
  /** @nullable */
  totalSeats?: number | null;
}

export interface CollegeCard {
  id: string;
  name: string;
  slug: string;
  location: string;
  state: string;
  /** IIT, NIT, IIIT, GFTI, etc. */
  type: string;
  established: number;
  /** @nullable */
  naacGrade: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  avgPackageLPA: number | null;
  /** @nullable */
  peakPackageCR: number | null;
  /** @nullable */
  cseCutoffRank: number | null;
  /** @nullable */
  nirfRank?: number | null;
  featured?: boolean;
}

export interface ImportantDate {
  event: string;
  date: string;
}

export interface CounsellingDetail {
  id: string;
  name: string;
  slug: string;
  type: string;
  level: string;
  states: string[];
  startMonth: string;
  endMonth: string;
  startYear: number;
  /** @nullable */
  logoUrl: string | null;
  description: string;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  totalColleges?: number | null;
  /** @nullable */
  totalSeats?: number | null;
  rounds: number;
  colleges: CollegeCard[];
  importantDates?: ImportantDate[];
}

export interface Round {
  id: string;
  counsellingId: string;
  roundNumber: number;
  name: string;
  registrationStart: string;
  registrationEnd: string;
  resultDate: string;
  year: number;
  /** @nullable */
  notes?: string | null;
}

export interface Lab {
  name: string;
  description: string;
}

export interface College {
  id: string;
  name: string;
  slug: string;
  location: string;
  state: string;
  type: string;
  established: number;
  /** @nullable */
  naacGrade: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  description: string;
  /** @nullable */
  about?: string | null;
  admissionThrough?: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  avgPackageLPA: number | null;
  /** @nullable */
  medianPackageLPA?: number | null;
  /** @nullable */
  peakPackageCR: number | null;
  /** @nullable */
  peakInternationalPackageCR?: number | null;
  /** @nullable */
  cseCutoffRank: number | null;
  /** @nullable */
  nirfRank?: number | null;
  /** @nullable */
  nirfRankCategory?: string | null;
  /** @nullable */
  qsWorldRank?: number | null;
  /** @nullable */
  placementPercent?: number | null;
  /** @nullable */
  totalOffers?: number | null;
  /** @nullable */
  linkedinAlumni?: number | null;
  /** @nullable */
  totalFaculty?: number | null;
  /** @nullable */
  totalStudents?: number | null;
  /** @nullable */
  campusAreaAcres?: number | null;
  counsellings?: string[];
  topRecruiters?: string[];
  facilities?: string[];
  labs?: Lab[];
  tags?: string[];
}

export interface SectorBreakdown {
  sector: string;
  firms: number;
  percentage: number;
}

export interface YearPlacement {
  year: number;
  avgPackageLPA: number;
  peakPackageCR: number;
  totalOffers: number;
  placementPercent: number;
}

export interface PlacementData {
  collegeId: string;
  year: number;
  totalOffers: number;
  companiesRegistered: number;
  companiesOffered: number;
  internationalOffers: number;
  placementPercent: number;
  avgPackageLPA: number;
  medianPackageLPA: number;
  peakPackageCR: number;
  /** @nullable */
  mbdPlacedPercent?: number | null;
  topRecruiters: string[];
  sectorBreakdown: SectorBreakdown[];
  yearWiseTrend?: YearPlacement[];
}

export interface Cutoff {
  id: string;
  collegeId: string;
  counsellingId: string;
  year: number;
  round: number;
  branch: string;
  /** @nullable */
  branchCode?: string | null;
  category: string;
  /** @nullable */
  quota?: string | null;
  openRank: number;
  closeRank: number;
  /** @nullable */
  gender?: string | null;
}

export interface Hostel {
  name: string;
  /** Boys or Girls */
  type: string;
  capacity: number;
  /** @nullable */
  feePerYear?: number | null;
  /** @nullable */
  floors?: number | null;
  amenities?: string[];
}

export interface HostelData {
  collegeId: string;
  totalHostels: number;
  boysHostels: number;
  girlsHostels: number;
  totalCapacity: number;
  /** @nullable */
  boysCapacity?: number | null;
  /** @nullable */
  girlsCapacity?: number | null;
  /** @nullable */
  feePerYear?: number | null;
  /** @nullable */
  description?: string | null;
  amenities?: string[];
  hostels?: Hostel[];
}

export interface Department {
  id: string;
  collegeId: string;
  name: string;
  code: string;
  seats: number;
  /** @nullable */
  avgPackageLPA: number | null;
  /** @nullable */
  peakPackageLPA?: number | null;
  /** @nullable */
  cseCutoffRank: number | null;
  duration?: string;
  /** @nullable */
  description?: string | null;
}

export interface InternshipData {
  collegeId: string;
  year: number;
  totalInterns: number;
  avgStipendPerMonth: number;
  /** @nullable */
  peakStipendPerMonth?: number | null;
  topCompanies: string[];
  internationalInternships: number;
  /** @nullable */
  conversionToFullTime?: number | null;
}

export interface PlatformStats {
  totalColleges: number;
  totalCounsellings: number;
  totalCutoffRecords: number;
  totalStates: number;
  lastUpdated: string;
  topIITs?: number;
  topNITs?: number;
}

export type ListCollegesParams = {
search?: string;
counsellingId?: string;
state?: string;
type?: string;
featured?: boolean;
};

export type GetCollegeCutoffsParams = {
counsellingId?: string;
round?: number;
year?: number;
branch?: string;
category?: string;
};